Remove v2 directory and move io.py to utils/io.py
