+++ src/coreason_manifest/utils/io.py
